server/financeserver/requirement.txt
django
djangorestframework
requests
yfinance
yahooquery
django-cors-headers
PyPortfolioOpt
pandas
torch
scikit-learn
matplotlib
PyPortfolioOpt
scikit-learn